requirements.txt

opencv-python
numpy
pillow
google-generativeai
python-dotenv
PyQt6